Rename platform files to match class names, tidy cmake configs
