+++ platforms/android/tangram/src/main/cpp/androidPlatform.h
