+++ platforms/ios/src/TangramMap/iosPlatform.h
