+++ platforms/osx/src/osxPlatform.mm
@@ -1,11 +1,11 @@
 #ifdef PLATFORM_OSX
 
-#import <cstdio>
-#import <cstdarg>
-
-#include "platform_osx.h"
+#include "osxPlatform.h"
 #include "gl/hardware.h"
 #include "log.h"
+
+#import <cstdarg>
+#import <cstdio>
 
 #define DEFAULT "fonts/NotoSans-Regular.ttf"
 #define FONT_AR "fonts/NotoNaskh-Regular.ttf"
