+++ tests/src/mockPlatform.h
