+++ toolchains/iOS.cmake
@@ -10,13 +10,11 @@ set(FRAMEWORKS CoreGraphics CoreFoundation QuartzCore UIKit OpenGLES Security CF
 message(STATUS "Linking with Tangram Framework " ${TANGRAM_FRAMEWORK})
 message(STATUS "Building for architectures " ${ARCH})
 
-# ios source files
-set(IOS_EXTENSIONS_FILES *.mm *.cpp *.m)
-foreach(_ext ${IOS_EXTENSIONS_FILES})
-    find_sources_and_include_directories(
-        ${PROJECT_SOURCE_DIR}/platforms/ios/demo/src/*.h
-        ${PROJECT_SOURCE_DIR}/platforms/ios/demo/src/${_ext})
-endforeach()
+set(SOURCES
+    ${PROJECT_SOURCE_DIR}/platforms/ios/demo/src/AppDelegate.m
+    ${PROJECT_SOURCE_DIR}/platforms/ios/demo/src/main.m
+    ${PROJECT_SOURCE_DIR}/platforms/ios/demo/src/MapViewController.m
+    )
 
 get_mapzen_api_key(MAPZEN_API_KEY)
 add_definitions(-DMAPZEN_API_KEY="${MAPZEN_API_KEY}")
@@ -26,7 +24,6 @@ configure_file(${PROJECT_SOURCE_DIR}/platforms/ios/demo/Config.plist.in
     ${PROJECT_SOURCE_DIR}/platforms/ios/demo/resources/Config.plist)
 
 add_bundle_resources(IOS_DEMO_RESOURCES "${PROJECT_SOURCE_DIR}/platforms/ios/demo/resources/" "Resources")
-file(GLOB_RECURSE IOS_DEMO_SOURCES ${PROJECT_SOURCE_DIR}/platforms/ios/demo/src/**)
 
 add_executable(${EXECUTABLE_NAME} MACOSX_BUNDLE ${HEADERS} ${SOURCES}
     ${RESOURCES} ${IOS_DEMO_RESOURCES} ${TANGRAM_FRAMEWORK})
