+++ toolchains/tizen.cmake
@@ -48,12 +48,15 @@ add_subdirectory(${PROJECT_SOURCE_DIR}/core)
 
 set(LIB_NAME tangram) # in order to have libtangram.so
 
-# add sources and include headers
-find_sources_and_include_directories(
-  ${PROJECT_SOURCE_DIR}/tizen/inc/*.h
-  ${PROJECT_SOURCE_DIR}/tizen/src/*.cpp)
+add_library(${LIB_NAME} SHARED
+  ${PROJECT_SOURCE_DIR}/platforms/tizen/src/platform_gl.cpp
+  ${PROJECT_SOURCE_DIR}/platforms/tizen/src/platform_tizen.cpp
+  ${PROJECT_SOURCE_DIR}/platforms/tizen/src/urlWorker.cpp
+  )
 
-add_library(${LIB_NAME} SHARED ${SOURCES})
+target_include_directories(${LIB_NAME} PUBLIC
+  ${PROJECT_SOURCE_DIR}/platforms/tizen/inc
+  )
 
 # link to the core library, forcing all symbols to be added
 # (whole-archive must be turned off after core so that lc++ symbols aren't duplicated)
